feat(sveltekit): Update default integration handling & deprecate `addOrUpdateIntegration` (#10263)
This updates the last usage of `addOrUpdateIntegration` and deprecates
it.
